Improve container-based builds and support mkinitcpio images
zbm-build.sh now tells the difference between old (pre-2.0) and new
repository layouts and will correctly set up its container to build in
either environment. The default buildroot has been moved to /build so it
is much easier to launch a build container with a single bind-mount that
contains a configuration as well as optional hostid and zpool.cache and
will hold a build/ subdirectory with build products afterwards.