Use column to layout footer
Realistically, we only need to print a footer in two modes:

1) using the explicit column layout designated by the caller - typically
   three columns with equal padding in columns

2) using a single column, due to a narrow screen

`column` supports both of these use cases by padding columnar data (as
indicated by embedded line returns in stdin).

item1:item2:item3
item4::item5

Would result in two rows with three columns. The second row would have
an empty middle column. By simply replacing ':' with a line return, we
can force everything into a single column.

`column` is not provided by BusyBox, but is part of util-linux. If the
binary is not found on the host system at build time it is internally
disabled via the `HAS_COLUMN` use flag. A subset of text is used on each
screen in that condition, or if the width is less than 80 characters.

Closes #254
Closes #256