Force creation of the finished hook at runtime
Setting rd.hostonly=0 in the release and recovery images causes dracut
to throw out the initqueue hooks that ZBM installs to force the event
loop to fire at least once. As a result, the console is often not
initialized before ZBM takes control.

Use a fake cmdline hook to create this for us after initqueue/finished/
has been destroyed.