Extend capabilities of zbm.prefer
Under certain circumstances, it might be preferable to ensure
ZFSBootMenu only ever interacts with a specific zpool.
Historical/default behavior has every pool being imported each boot,
with optional preference being given to one specific pool.

A new modifier to `zbm.prefer`, `!!` has been added. When this is
appended to the name of a pool, that pool is imported exclusive of all
others on the system. If the pool can't be imported, a retry/recovery
shell loop is initiated.