Move dataset mountpoint to $BASE/environments/
Creating a directory tree for zpools / datasets rooted in $BASE could
potentially lead to a conflict between one of the various files we
create and a pool name. Moving the root of this to $BASE/environments/
means that there can't be a conflict, since imported pool names must be
unique.

Additionally, a helper was added to return a string with the path to a
given filesystem on disk. This avoids manually constructing the path and
having to update those individually should this directory location ever
change.