Add/use border label feature flag
fzf 0.35 added the ability to added labels to borders - both for the
entire application window and for the preview window. This can now be
used to add a global header, present on all fzf instances. The snapshot
screen can now optionally move context text to the bottom border around
the preview window instead of being in it.