Key management: properly normalize Booleans, make cache mutex per-PID
The cache mutex exists to avoid infinite recursion when attempting to
load a key for an encryption root from its key source. However, the
recursion is only a concern when the load is done in the same PID that
is trying to do the caching. In other cases, such as when a remote user
has terminated a local instance and is not attempting to unlock a pool,
a generic mutex will prevent all key caching. Making the mutex per-PID
avoids this conflict.