zfsbootmenu: manage efivarfs when needed
initcpio mounts efivarfs as read-write, dracut doesn't mount it at all.
Inside zfsbootmenu-init, we now check if the system is in EFI mode and
then mount or remount efivarfs as read-only.

When chrooting into a boot environment, efivarfs is mounted in the
chroot as read-only if the zpool is read-only, and read-write if the
zpool is read-write.

When entering a recovery shell, efivarfs is re-mounted as read-write if
efibootmgr is present in the initramfs. Exiting the shell remounts it as
read-only.