zsnapshots: recovery environment helper tool
In the event that no boot environments with a valid /boot are found,
recovery might be as simple as interacting with an older snapshot. The
recovery shell tool `zsnapshots` attempts to make that easier by allowing
access to the main snapshot interface for a given dataset.

fzf defaults suitable for the main interfaces were moved to a lib file
that can be sourced. These defaults allow smaller tools to match the
same behavior that fzf has when invoked from /bin/zfsbootmenu .

select_kernel now returns 1 when the 'kernels' file for a boot
environment can't be found, with nothing printed. The caller is expected
to handle that condition accordingly.

Fixes #418