zfsbootmenu: protect stdout of main interface
fzf forces us to capture stdout to know which key combination was
pressed and which item was selected. A bug has been lurking for ages:

* Hit mod-l on the main screen to see the warning/error log
* Hit enter to select whatever is shown
* Hit mod-p to look at the pool status screen

When the sub fzf instance used to draw logs exits with a selection (e.g.
with the enter key), it writes that to stdout. Then when the lower level
fzf instance exits (e.g. the main menu screen) with a selection, stdout
from both instances is captured by the subshell.

This solution protects against that by discarding stdout from all of the
sub-instances of fzf.