zbm-builder.sh: Add support for running on hosts with SELinux enabled
This commit updates `zbm-builder.sh` to include a new `-M` option for
specifying podman/docker volume mount flags. On systems with SELinux
enabled, `-M z` must be specified so that the volume will be relabeled
with the `container_file_t` SELinux type. Otherwise, the directory
contents are not accessible within the container.

In addition, when using `-l <source dir>`, `-O -e=DRACUT_NO_XATTR=1`
must be specified. By default, when dracut copies files to its
temporary directory, it runs:

    cp --preserve=mode,xattr,timestamps,ownership <...>

This environment variable makes dracut disable the `xattr` flag. This is
necessary because cp would try to preserve the `security.selinux` xattr,
which is never allowed inside a container. This should have no impact on
the generated initramfs image because ZFSBootMenu does not use any
custom xattrs and cpio is incapable of storing them anyway. mkinitcpio
does not try to set xattrs, so no workaround is needed for that.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>