zbm-builder.sh: add mount-options argument, document use
A new `-M` option for `zbm-builder.sh` provides a means to specify mount
options for the `/build` and (if specified) `/zbm` volumes in the build
container. Using, for example, `-M z` allows images to be built on hosts
with SELinux enabled.

This change is documented, along with extra considerations for building
Dracut-based images in containers when SELinux is enabled on the host.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>

Closes: #433.