zfsbootmenu-core: try to load keys in mount_zfs
Attempting to load encryption keys when mounting a ZFS filesystem avoids
the need to manually load keys in a few places and ensure that mount_zfs
tries hard to do the right thing even when run interactively.

Because mount_zfs is also called from cache_key and it might be
problematic to enter some indirect recursion when manipulating ZFS keys,
setting NO_LOAD_KEY=1 will tell mount_zfs to skip this attempt and just
fail when the BE is locked.