zfsbootmenu-lib: show sort key where needed
When fzf supports a border label (>= 0.35.0), show the current sort key
that is being used to sort the list. This is only relevant on the main
BE selection screen and the Snapshot selection screen. If fzf does not
support a border label, fall back to legacy behavior of not showing the
value anywhere.