zfsbootmenu-core: disable spl_hostid warnings
The default behavior of hostid matching is useful to have on in binary
releases. However, it generates a warning message that isn't really
relevant for users of that build.

It's expected that the spl_hostid value in ZFSBootMenu.EFI won't match
the local pool hostid - advising these users to manually set `spl_hostid`
on the ZBM KCL runs counter to how we recommend ZBM should be used.