zfsbootmenu: remove legacy root= usage/design
ZFSBootMenu has danced around `$root` for a few years now, as a holdover from
very early design decisions picked up from Dracut. We don't actually _need_ to
tie ourselves to that variable - and in fact, a lot of the early startup process
can be simplified by no longer trying to stuff so much logic into it.

To that end, the legacy but undocumented support for
`root=zfsbootmenu:POOL=pool` has been completely removed. `root` on the ZBM KCL
is no longer honored. `zbm.prefer` with the documented syntax is now the only
way to tell ZFSBootMenu which pool should be preferentially used in the boot
process. Accordingly, the global `zbm_prefer_pool` variable is now set when a
pool should be preferred, and the accompanying `zbm_require_pool` variable is
used to indicate how firm that pool requirement is. One quirk of this is that a
pool name of `zfsbootmenu` can now actually be a value that `zbm.prefer` will
respect/honor.