bin/generate-zbm: require at least one kernel version value
Some distributions create generic symlinks to a kernel; such as vmlinuz.old.
These symlinks do not encode any version information in the filename. On ARM
systems, the kernel sometimes does not contain a version that can be detected
via dumping strings from the file. This leads to a case where neither version
string is defined, and the assumption was that one would always be defined.