aarch64 support: efi stub and devicetree selection
generate-zbm now understands the EFI.DeviceTree key. This key can
contain one of three possible values:

1) /absolute/path/%{kernel}/to/device-tree-file.dtb
2) /absolute/path/to/device-tree-file.dtb
3) partial/path/device-tree-file.dtb

If found, %{kernel} is replaced with the kernel version being built in
to the EFI asset. Partial paths are appended to /boot/dtbs/dtb-$kernel/,
a well-known path for device tree files used by Void Linux.

Inside ZFSBootMenu, the org.zfsbootmenu:devicetree property is read. The
value of this property are handled in the same order/priority as
EFI.DeviceTree. %{kernel} is replaced with the kernel version that kexec
is loading / executing.